fix(ballot-interpreter): reject corners we are not sure about (#5220)
Closes https://github.com/votingworks/vxsuite/issues/4913.

We were not being very careful about ensuring the edges first/last timing marks agreed and were thus quite likely to be corners. However, naively ensuring that we always found the corner timing marks rejected a lot of ballots due to the shape of those marks sometimes being incorrect. This might occur when the corners are quite close to the edge.

To address this, I re-used the logic to determine which of the bottom timing marks are present/absent and expanded it to look for white surrounding the black of the timing mark itself. This ensures that images from when a ballot corner is folded down are properly rejected, but minor scanning artifacts do not interfere with inferring the corner timing marks.