VxScan: Ignore jam flag if ballot clears rear sensors on reject
When a double sheet is detected and the ballots are stopped midway
through scanning, they are sometimes stopped very quickly and remain
close to the front of the scanner. In these cases, ejecting them to the
front can cause them to shoot out of the front of the scanner and then
slide back down the infeed. The PDI scanner has a `documentJam` flag set
after this, which currently leads to us showing a jam message
instead of a double sheet message.

To fix this, we ignore the `documentJam` flag in the case that the
ballot successfully cleared the rear sensors on eject, since an actual
jam would likely leave the ballot covering both sensors.