VxScan: Handle double sheet error when scan doesn't fail
Originally, when building double sheet detection, I saw a `scanFailed`
event after every `doubleSheetDetected` event. Now, it seems like there
are some cases where we don't get a `scanFailed` event and instead get a
`scanComplete` event. In these cases, the ballots are still stopped
partyway through the scanner. In addition, the `documentJam` flag is
set.

The state machine was doing an ok job at handling this, but only because
the status after the `scanComplete` event contained the `documentJam`
flag. Instead of relying on that flag, let's be more precise and
actually handle the double sheet error after a `scanComplete` event
explicitly.