VxScan: Don't error if ballot removed but scan completes (#5225)
Previously, if there was no ballot in the scanner after a scan
completed, we would declare an error and try to disconnect/reconnect to
the scanner, since this seemed like an impossible case. It turns out you
can produce this case by pulling the ballot out soon after the scan
begins. Sometimes this will result in a failed scan, but other times the
scan will complete even though the ballot is gone. Since we're confident
in that case that there's no ballot in the scanner, we can abandon the
scan and wait for a new ballot to be inserted. This is an improvement
because we don't have to spend time disconnecting/reconnecting (and we
have seen at least sometimes that disconnecting and reconnecting fails,
which then requires a restart of the entire app).