feat(converter-nh-accuvote): write printing and proofing ballots (#5267)
When converting we now write out two ballot PDFs for each ballot style/precinct combination. The first is prefixed with PRINT and has the QR code to allow it to be scanned, the second overlays information about where the bubbles and contest options are encoded in the election definition. This allows one to quickly determine whether the generated election definition is correct.