Increase max election file size to 30MB
During scale testing, this limit prevented us from testing with a large
mock election file. I temporarily increased it to enable the testing,
and the system was generally able to handle a large election around 30MB
without issue. So I'm increasing this limit to the tested size so that
it doesn't end up accidentally biting us later in the extremely rare
case that we have an election over 10MB.