VxDesign: Update vendored translation for Spanish ballot label (#5350)
The ballot labels for the different ballot variants (sample/test/official and
precinct/absentee) all need to fit on one line in order for ballot
layout to work. In Spanish, the most verbose language we support, some
of the auto-translations are too long, so we use shorter vendored
translations instead. These recently got reset, so updating them again
to make ballot layout work.