VxAdmin: Disable PDF preview/export for large reports (#5343)
* printing: Limit content size for printing to PDF

Chromium crashes with an out of memory error when trying to generate a
PDF for very large documents. Set a limit well below that point to
prevent crashes and return a usable error code.

* admin/backend: Propagate PDF errors to frontend for previews

We'll use the preview error to disable the buttons to print/export PDF
reports, so we don't need to propagate those errors.

* admin/frontend: Show warning when PDF is too large and disable export

* admin/frontend: Disable PDF previews for long PDFs

At around 100 pages, the browser starts crashing occasionally during
scrolling. To prevent this, set a conservative limit of 50 pages.

* Don't render empty reports

Previously, reports with no data showed up as a blank page. Instead, we
don't render them at all.

* PDF viewer: Reset page count when PDF data changes

In the report builder, changing the parameters would remove the current
PDF preview but leave the page count.

* Update other `renderToPdf` calls to throw content-too-long errors

We don't expect any of these PDFs to reach the length limit. If they do,
it's an unexpected error, so we can crash explicitly, since that's an
improvement over hanging or crashing with a Chromium error.

* Update tests

* Improve failure logging for report previews