Add UiRichTextString component
The react-i18next's simple HTML tag preservation feature doesn't work
for `img` tags - it strips all of the attrs. The `Trans` component has
support for a more complex tag replacement feature, where the tags are
replaced with indexed placehodlers in the translation strings, but I
couldn't get it working.

The basic react-i18next translation function (which just substitutes in
the correct translated string) doesn't seem to mess with the HTML tags
at all, so I used that instead of the `Trans` component.  I reasoned
that somewhere under the hood, `Trans` was injecting the HTML string
into the DOM, so I might as well just do that explicitly and sanitize it
explicitly.

The downside is that now we have a separate component for translating
rich text, but overall I think things are clearer and more explicit, so
I'd consider it an overall improvement.