Replace DOMPurify with sanitize-html
Standardize on the sanitize-html library for XSS protection, since it's
written in pure JS and thus can be run on the frontend or backend.
DOMPurify utilized the DOM and thus could only be run in a browser (or
needed jsdom on the server).

It turned out that most of the places we were using DOMPurify were
unused legacy code anyway, so I removed those call sites as well.