style: use `node:` scheme
Adds `eslint-plugin-n` and enables the rule to require importing nodejs built-in modules with the `node:` prefix. This ensures we don't accidentally load a package from NPM when we expected a builtin. It also makes our imports consistent with future modules that are only availble with the prefix such as the `node:test` module (https://nodejs.org/api/test.html).