Use `node:` import scheme for built-in NodeJS modules (#5411)
* fix: run stylelint fix script properly

* style: use `node:` scheme

Adds `eslint-plugin-n` and enables the rule to require importing nodejs built-in modules with the `node:` prefix. This ensures we don't accidentally load a package from NPM when we expected a builtin. It also makes our imports consistent with future modules that are only availble with the prefix such as the `node:test` module (https://nodejs.org/api/test.html).

* style: update generated code to use `node:` scheme

* style: update non-automatic imports of node modules

* test: sort before comparing fs results

Reading directory entries does not have a guaranteed order.

* build: add more config for handling `node:`-prefixed imports

* style: convert a few `requireActual` calls to use `node:` prefix