ballot-interpreter: Crop black borders in diagnostic
To ensure that the diagnostic works with precinct scanner images, apply
the same cropping logic that we use when interpreting a ballot. Note
that we still keep the existing logic that skips over a margin when
checking the ballot - this is useful in case there's any skew that
leaves some black pixels that don't get cropped.