Use pulse audio commands instead for switching output ports.
Also only attempt the commands for NODE_ENV="production", since they
won't work in other environments.

Lastly, explicitly throwing exceptions if the commands fail, as a last
resort to end the session if we can't play the alarm sound over the
speaker (or if we can't switch back to headphone audio).