VxAdmin: Only show card programming modal on smart cards screen
In addition, add the smart cards screen to the system admin left nav
when unconfigured (to maintain the existing ability to program sys admin
cards without configuring the machine). Note that currently the smart
cards screen will have some misleading copy about creating election
cards when unconfigured, but I'll be updating that screen in the next
PR, so it's just a momentary issue.