chore(deps): update `typescript` to v5.6.2
Changes in v5.6: https://devblogs.microsoft.com/typescript/announcing-typescript-5-6/

I don't see anything that dramatically affects us. The change in `assert.test.ts` is due to the "Disallowed Nullish and Truthy Checks" change. Once we update Electron and NodeJS to versions that support it, the "Iterator Helper Methods" changes will remove some of the need for `iter` in `libs/basics`, but we can't use them yet.