VxAdmin: Allow continuous CVR exports to be re-imported (#5443)
Previously, if you imported a continuous CVR export from VxScan into VxAdmin, then scanned more ballots and tried to re-import the same export, the "Load" button would be disabled. This fixes the logic for disabling that button to check the export timestamp, which changes when new ballots are exported (instead of just checking the file name, which doesn't change). The backend already knows how to only import the newly added CVRs, so the only fix needed is enabling the button.