VxAdmin/VxCentralScan: Wrap long screen titles (#5444)
Instead of letting the buttons or other content in the header wrap internally, wrap the titles. This is currently only an issue on the Voting Method Ballot Count Report screen, but is good to have in place for the future. There might be a larger redesign we can do so that the actions don't take up as much header space, but this is a viable quick fix, and since the issue only affects one screen, it's probably good enough for now.