VxScan: Don't try to recover from unexpected errors during scanning (#5465)
* scan/backend: Fail fast on unexpected PDI/interpretation errors

Don't reject the ballot, to keep things simple. Also, don't try to recover, since that will also mask errors.

* Ensure 100% coverage for PDI state machine

That way we know we're testing every error case

* ballot-interpreter: Crash on malformed JSON from Rust addon

This is an unexpected error, not an unknown error. Previously would
result in an unreadable ballot instead of a crash.

* Cleanup: rename outdated type

From when ballot-interpreter was ballot-interpreter-next so we called
everything related to it "next"