fix(ballot-interpreter): improve vertical streak detection (#5522)
* fix(ballot-interpreter): exclude more ballot edges

When scanning in VxCentralScan sometimes we get wider gray areas outside the ballot paper than we expected. Most of this gray is binarized to black, leading to false positives when detecting streaks. This reduces the likelihood of false positives while preserving detection of streaks within the ballot, including within the timing marks.

* feat(ballot-interpreter): improve vertical streak debug images

Binarizes the debug image to make it clearer what the streak detector was working with.

* test(ballot-interpreter): fix comment ⟷ code mismatch

The comment says to draw on side B, so I updated the code to do that.

* test(ballot-interpreter): add more streak detection tests

Ensures that streaks through timing marks are still detected and semi-wide edge "streaks" do not cause false positives.