Add vertical streak detection controls (#5549)
* fix(ballot-interpreter): correct JS argument marshaling

If `options` was given as `{ scoreWriteIns: undefined }`, the interpreter would fail to coerce the value as a `bool` but would ignore the error until some point in the future when interacting with the JavaScript context. This is because the `get::<JsBoolean>` call would put the context into a throwing state, but would not then bubble that `Throw` up to the caller. As a result, subsequent calls to the context would also yield `Throw` for operations that should have succeeded. In this case the error would be "failed to downcast any to boolean".

To fix this we need to not trigger throws when we don't intend to bubble them up. In particular, instead of `get::<JsBoolean>` we now `get_value()?.downcast::<JsBoolean>` so that we can bubble possible errors from the getter but then handle a mismatched type in Rust appopriately.

* feat(scan): allow disabling streak detection

Closes #5538

Adds a system setting that disables streak detection for an election. We might want this in case of a ballot design that triggers false positives when looking for streaks. Because we support taking someone else's AccuVote-style ballot, we may not be able to ensure no streaks are naturally present in the original image. This is unlikely, but this gives us an escape hatch.

* refactor(scan): access system settings directly

* fix(ballot-interpreter): pass `disableVerticalStreakDetection`

Ensures that the `disableVerticalStreakDetection` system setting is passed through from the CLI. Also adds a CLI option to disable streak detection.

* test(scan): ensure disabling streak detection works