fix(central-scan): prevent `scanimage` race
Ever since #4994 we call `scanimage` twice when starting a batch, once to determine whether an imprinter is present and once to actually start the scan. Both invocations require communicating with the scanner, which means both need to open the USB device. Because we previously resolved the `isImprinterAttached` promise early in the case where the expected error message is present, the first `scanimage` instance could still be running and holding the USB device open when we try to run the second `scanimage` instance. When that happens, the second instance would fail to open the USB device and we'd get a zero-sheet batch and this error message: `scanimage: open of device fujitsu failed: Invalid argument`.

In my testing on macOS host with a Debian VM with the attached fi-7180, this happens on every scan attempt. It may be that different configurations behave differently, as the outcome of the race is non-deterministic. This change fixes the issue by waiting until the `scanimage` process exits to resolve the `isImprinterAttached` promise.

Additionally, I changed a few other things:
- expand the shorthand `scanimage` option `-d` to `--device-name`
- collect `scanimage` stderr before checking for the expected error message in case it is split across multiple chunks (unlikely but possible)
- update tests to fail if the `isImprinterAttached` promise resolves early or if we don't see the error message split across multiple stderr chunks