fix(central-scan): delete temporary batch holding directory
When scanning ballots in VxCentralScan we use `scanimage`, which puts images on disk. We then read them, interpret them, normalize them, and write them to disk at a new location of our choosing. This second image is the only one we really care about, and the first is just taking up space. To fix this, we now delete the directory we create for batches after the batch is completely scanned.

Additionally, this commit removes the `Workspace::scannedImagesPath` property and stops creating the `scanned-images` directory in the workspace. Before we put files for a given batch in their own directory we would put them in this directory. That hasn't been true in some time.