fix(central-scan): prevent `scanimage` race (#5570)
* fix(central-scan): prevent `scanimage` race

Ever since #4994 we call `scanimage` twice when starting a batch, once to determine whether an imprinter is present and once to actually start the scan. Both invocations require communicating with the scanner, which means both need to open the USB device. Because we previously resolved the `isImprinterAttached` promise early in the case where the expected error message is present, the first `scanimage` instance could still be running and holding the USB device open when we try to run the second `scanimage` instance. When that happens, the second instance would fail to open the USB device and we'd get a zero-sheet batch and this error message: `scanimage: open of device fujitsu failed: Invalid argument`.

In my testing on macOS host with a Debian VM with the attached fi-7180, this happens on every scan attempt. It may be that different configurations behave differently, as the outcome of the race is non-deterministic. This change fixes the issue by waiting until the `scanimage` process exits to resolve the `isImprinterAttached` promise.

Additionally, I changed a few other things:
- expand the shorthand `scanimage` option `-d` to `--device-name`
- collect `scanimage` stderr before checking for the expected error message in case it is split across multiple chunks (unlikely but possible)
- update tests to fail if the `isImprinterAttached` promise resolves early or if we don't see the error message split across multiple stderr chunks

* fix(central-scan): only keep two images per scanned sheet (#5571)

* docs(central-scan): fix typo

* refactor(central-scan): wrap batch data in an object

This way we don't have a situation where one of the fields might be set and the other isn't.

* fix(central-scan): delete temporary batch holding directory

When scanning ballots in VxCentralScan we use `scanimage`, which puts images on disk. We then read them, interpret them, normalize them, and write them to disk at a new location of our choosing. This second image is the only one we really care about, and the first is just taking up space. To fix this, we now delete the directory we create for batches after the batch is completely scanned.

Additionally, this commit removes the `Workspace::scannedImagesPath` property and stops creating the `scanned-images` directory in the workspace. Before we put files for a given batch in their own directory we would put them in this directory. That hasn't been true in some time.