fix(scan): improve acceptance rate of skewed ballots
When a ballot has a significant skew (>= 1 degree), the previous implementation of `find_scanned_document_inset` would cause too much of the ballot to be cropped out of the image. Because it would only stop cropping once the white pixels were half of a given row, the portion of a skewed ballot that would be cropped was likely to contain timing marks. At the very least, it was likely to partially or completely crop the corner timing mark, which we require in order to proceed with interpretation.

The new implementation takes a configurable ratio of white to black pixels at which to stop cropping, which we now set to 10% (vs the previous 50%). This makes the cropping much more conservative while still cropping off any edge rows/columns of pure black. This preserves the corner timing marks even with skew around 3 degrees, which is as high as we're likely to be able to interpret successfully.