fix(scan): allow more cropping of bottom corners
On ballots with timing marks printed right up close to the edge, scanning sometimes crops the edge timing marks in a way that we were too aggressively discarding. This change enforces that interior timing marks still be substantially uncropped, but allows the edges (i.e. corners) to be down to 1/3 of their original size.

Also adds a new debug image which I used to help me diagnose the problem.