fix(scan): improve ballot interpreter skew handling (#5633)
Changes the approach to identifying contours as timing marks by using a technique similar to a template image match to compute scores for the timing mark's "mark" and "padding" parts, then using those scores to both filter out anything too low and to select sets of timing marks along the borders that maximize those scores.

This fixes two issues identified by the TRR corpus:
1. Though we had all the timing marks, the ones in the region of the image that was non-trivially skewed exhibited rotation that increased their height, so the initial filter of contours would reject them. We would then later infer their location, but inferring side timing marks triggers skew/rotation limiting that caused the ballots to be rejected.
2. Sometimes there is debris around the edges of the ballot that ends up being identified as possible timing marks. When this happens, we would sometimes select one of these fake marks instead of a real one when finding the set of marks that make up a border. This is because we only cared about finding the correct number of marks, but didn't do anything to measure their quality. Now we use the computed scores mentioned earlier to break ties when both sets of timing marks have equally correct counts.

Includes a faster version of the timing mark border finder that prioritizes likely winners and short circuits when it finds an exact match. If it can't find the border it's looking for, it falls back to a slightly edited version of the previous slow path.