Fix bug in contest layout for HMPBs (#5639)
When laying out multiple contest sections on the same page, the code
calculates whether there is enough vertical space left before adding a
new section. That calculation forgot to include the gap between
sections. This caused it to believe there was more space available than
there actually was. Ocassionally, this would result in adding more
contests than would fit on the page, causing the page content to grow a
bit longer and eat into the bottom margin of the page.