dev-dock: Parameterize printer mock config rather than machine type
Instead of passing in the machine type (i.e. `scan` or `admin`) to the
dev dock and then conditioning internally to determine the printer
mocks, pass the printer config in as a parameter.

This change set up a pattern for passing in more machine-specific values
to the dev dock (e.g. a handler for the PDI scanner mock).