refactor(res-to-ts): convert package to ESM
This doesn't really have a lot of effect in this package since we don't actually build it for use in other packages. But in order to move any package to useESM (i.e. `type=module`) we need all of the workspace packages that it imports to be ESM too. This means we need to start with the packages that nothing imports directly.