feat(test-utils): work with either `jest` or `vitest`
Follow-up to https://github.com/votingworks/vxsuite/pull/5690.

Changes `@votingworks/test-utils` to look for either a `jest` global or import from `vitest`, mostly for creating mock functions (i.e. `fn`). For simplicity, keeps the `jest` API for mock functions because they're so similar (but not actually TS-compatible). The possibility of `import`ing whenever we call `fn` within `test-utils` means that anything that creates a momck function now needs to be async. This isn't awesome and it caused a bunch of changes in other packages, but I think it's worth it to allow the `jest` → `vitest` migration to occur in pieces.