test(backend): replace `jest` with `vitest` (#5773)
* test(backend): replace `jest` with `vitest`

This library provides various test utilities to the app backends, which are still using jest. So this change makes those utilities work with either vitest or jest in a similar way to what we've done before. The main difference this time is that we needed to pass global hooks/helpers through to some helpers, which can be a little tricky typing-wise. To help with that, I'm importing `@jest/globals` on both sides of the call to ensure we get the same types and that everything is happy.

Additionally, the mock Google service classes were changed to functions because TS overloading of class constructors is very much not fun.

* test(mark-scan): adjust thresholds

* test(ui): update snapshots

* test(mark-scan): fix mock setup

`scanAndSave` does call `getPaperHandlerStatus`, but it passes the result to `isPaperAnywhere` which this test suite mocks. It's not really clear to me whether this worked before, but the correct action we want to take happens when `isPaperAnywhere` returns false, so that's what we mock.