feat(test-utils): add `disposable` helpers (#5817)
Adds `disposable` and `asyncDisposable` for wrapping resources that have explicit cleanup functions that do not run automatically. Uses a new feature of JavaScript: the `using` and `await using` declaration kinds:

```ts
test('example', () => {
  using resource = disposable(createResource(), (r) => r.close());
 // Use `resource` here.
});
```

Most resources in the NodeJS ecosystem don't support this feature yet, so this is a way of adding it specifically for use in tests where we might otherwise fail to clean up a resource because of an expectation failure.