test(mark/frontend): replace `jest` with `vitest`
Closes #5835
Part of #5706

The main work involved was to do with fake timers. Using `vi.waitFor` is needed because `screen.findBy*` ends up using the fake timers and just times out. Additionally, `vi.waitFor` advances the fake clock by 50ms for each time it checks. This mostly had no effect, but in one test we expected the fake clock not to advance.