Find precinct split from ballot style in ballot preview
If a precinct has splits, VxDesign creates a separate ballot style for
each split. When previewing a particular ballot style, we know the
precinct ID and the ballot style ID, so we can figure out the relevant split ID
 (if there is one) from that information. If the precinct has splits,
then we look up which split this ballot style corresponds to.