VxDesign: Show expected errors during ballot preview (#5925)
Currently, there's just one type of expected error: when a contest is
too tall to fit on the page.

To deliver this error to the frontend during ballot proofing, we catch the error within the ballot template so we know what contest was too long and pass it up the stack using a `Result`.