fix(design): prevent crash on contest delete
Fixes https://votingworks.sentry.io/issues/6266281369/

### Steps to Reproduce
1. Create a contest.
2. Delete the contest.

### Expected Result
The contest should be deleted and the user should be redirected back to the contest list.

### Actual Result
Most of the time, it happens exactly as expected. However, the order in which `savedElection` is updated and the `updateElectionMutation` `onSuccess` callback is called is not defined. If `savedElection` is updated first it will no longer have the contest that was just deleted and the `find` call will panic.