test(design): re-enable `BallotsScreen` test
Changes the `setBallotsFinalizedAt` API to be `finalizeBallots` instead to avoid issues with the timestamps not quite matching due to the way vitest handles fake timers. Specifically, Jest freezes the current time to whatever it is and advancing time doesn't change it. In vitest, advancing time does alter the current time so we can't just use the same timestamp everywhere. Not taking a timestamp in the API means we don't have to worry about what it might be set to and can stop using fake timers in this test.