fix(design): correct some `ContestsScreen` UX issues
Re-enabling and fixing the tests revealed a few problems with this screen:
- deleting the First Name field would instantly insert the rest of the name as the first name; this was due to the "fallback" logic we had
- submitting without a District selected was allowed but produced an invalid contest

To fix the name issue, I changed the form to work with "draft" versions of the contest and candidate data structures that better fit the form. We then validate the data on form submit, both using the HTML `required` attribute and then using our Zod schemas.