test(design): fix broken check for image rendering
The check for the signature image was broken because it was finding the rendered seal image since they used the same dummy image. This was a problem because the tests waited until this image was found to click "Save" on the form to ensure that the signature image would be included in the updated precinct information. However, finding the seal image instead meant that clicking "Save" would sometimes happen early, making the update happen before the signature image was properly included with the request payload.