ci: perform recursive builds sequentially (#6045)
Something with `tsc` gets tripped up when multiple processes try to build the same package simultaneously. We haven't seen this be a problem in dev, so I'm only making this change in CI. It will slow CI down a bit, but it should reduce flakiness. Long term, I don't think we want this recursive `pnpm` package-based build setup, so I don't feel bad disabling it here.