fix(usb): detach kernel driver
While we shouldn't need this in production, and realistically won't need it with most USB devices, it's possible for the kernel to decide that it has a driver to handle one of our devices. In that event, it will open and claim the interface and trying to claim the interface ourselves will fail. This happens on a default install of Debian 12 with the Fujitsu thermal printer.