feat(scan): allow configuring scanner's bitonal threshold (#6034)
Adds a system setting called "Bitonal Threshold" which will be passed to the PDI scanner for black & white scans to determine the threshold for determining whether a pixel is black or white. Note that this setting cannot be per ballot style but must be per election (or theoretically, at highest granularity, per scanner). If ballots for an election are printed on multiple colors of paper then this single value must work appropriately for all of them. This likely means non-white color should be quite light if there are any white ballots in the mix and that the contrast between any two ballot paper colors should not be very high.

Though this will allow us to scan paper that is not white, scanning in grayscale would be a better way to go in the long run assuming there are no major tradeoffs to it. Scanning in grayscale means we can use a dynamic threshold, e.g. by applying otsu's method as we do in v3.