VxDesign: Return minimal info in listElections response (#6113)
* Return minimal info in listElections response

Don't send the entire election record to the frontend, just the fields
needed. Compute status on the backend. This isn't strictily necessary
yet, but sets up a more reasonable interface for easy
refactoring/optimizing of data loading from the db later. The basic goal
was just to remove all knowledge of ElectionRecord from the frontend.

* Backend tests

* Frontend tests