feat(scan): allow pausing clown car app components (#6143)
* test(scan): stop `Store` using real loggers

It just clutters stdout.

* fix(scan): actually exit `pdictl`

Without this, the `pdictl` processes pile up and prevent whatever the
currernt one is from connecting to the scanner.

* feat(scan): allow pausing clown car app components

Rather than having a single "stop test" button, we now have pause
buttons for each of the components that we're testing. Additionally, I
added a "Power Off" button to make safe shutdowns easier.

* chore(scan): remove Mario

* refactor(scan): rename `loop` to `task`

I had an earlier revision of `TaskController` that was called
`LoopController`.

* refactor(scan): update a few names for clarity