VxDesign: Relational db tables for election content (#6117)
* Migrate election content to relational tables

* Remove `ifNotExists` from table creation in migration

Migration should fail if tables already exist, since it means the db was
left in a weird state.

* Add primary keys to association tables

* Quote strings in migration

We don't get to leverage the built-in PG client quoting with these
commands, so we have to do it ourselves.

* Ensure candidate names are migrated to first/middle/last

* Move ElectionListing and ElectionStatus to types.ts

* Update `listElections` API method

* Update store.createElection

* Update store.getElection

I chose to keep the `getElection` method around for now instead of
creating different methods for each entity for a few reasons:
- Trying to keep this change as minimal as possible
- There are some other consumers of this method besides the API methods
  for each entity (e.g. rendering ballots, exporting election
definition)
- It should be easy to factor out into smaller methods in a subsequent
  PR

* Add store.updateElectionInfo

Replace the calls to `updateElection` and `updateBallotLanguageCodes` in the API method

* Update district store methods

Now we can rely on the database to maintain referential integrity instead
of manually updating references

* Update precinct store methods

Given the complexity of the precinct data model, for updating I thought
it would be simpler to delete and re-insert the precincts and their
splits (preserving IDs). Open to feedback on this.

* Update party store methods

* Update contest store methods

The tricky part of this is maintaining the correct ballot order.

* Update store.setBallotPaperSize and remove store.updateElection

Now that we've updated all the consumers of updateElection, we can
finally remove it.

* Split candidate names into first/middle/last when loading an election

* Order by created_at to preserve original order

For districts->precincts, precincts->precinct splits, and
candidates->parties, order by the related entity's ID to make a
deterministic order, since the actual order of the associations doesn't
matter.

* Update tests

* Use generateId function for precinct split IDs

* Clean up transaction handling

* Order districts/precincts/splits/parties alphabetically by name

* Use serial for contests.ballot_order

For query simplicity, better atomicity, and to match candidates.ballot_order

* Preserve ballot order when updating contest