Log Fujitsu printer getStatus error messages
We generally treat errors when getting the status as a sign that the
printer is disconnected. We can usually reconnect once the printer is
connected again by trying to reintialize the driver. However, we saw one
case where that didn't work, but we didn't have the original error
message to help debug. This change ensures that we log the original
error message so that we can investigate outlier cases.