VxMark: Show precinct/split name instead of ballot style ID (#6316)
* Add some utils for working with precincts/splits/ballot styles

* mark-scan: Poll worker activates precinct/split names instead of ballot styles

* mark-flow-ui: Inline ElectionInfo component in StartPage

It was only being used here. It handled more cases than was
necesssary for various undefined props, so removed that handling as
well.

* mark-flow-ui: Update StartPage to show precinct or split name

Don't show the ballot style ID anymore

* Tests for utils

* mark-flow-ui: Fix coverage ignore comments

* mark-scan: Update tests

* mark: Update tests impacted by start screen changes from mark-flow-ui

Note that I'm not doing the same full change in apps/mark yet, since I'm
not sure if it's important.

* Clean up precinct select

- No need to provide a value prop
- The change handler never receives undefined (manually tested), so
  add an assertion

* Re-add empty string value to select to make tests pass